Add Video: CI/CD pipeline with Docker, GitHub Actions, Node.js and AWS ECS Fargate
Hi,

I'm the CTO and Co-Founder of datree.io and AWS Community Hero. This is a hands-on Webinar with live coding that I'm presenting about how to build a CI/CD pipeline with Docker, GitHub Actions, Node.js and AWS ECS Fargate. We have hundreds of RSVPs and I genuinely think that it's valuable to this community.

https://dtr.ee/2DURIVu

Let me know what you think!

Best regards,
Shimon Tolts